requirements.txt
pandas
addict
sklearn
opencv-python
pytorch_warmup
scikit-image
tqdm
p_tqdm
imgaug>=0.4.0
Shapely==1.7.0
ujson==1.35
yapf
pathspec
timm
albumentations==0.4.6
pathspec
ptflops
